% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hest.sim.R
\name{hest.sim}
\alias{hest.sim}
\title{Basic Forward Simulation}
\usage{
hest.sim(model, init.data, params, dt, dt.sim, N, burn = 0, nreps = 1,
  max.bad.draws = 5000, verbose = TRUE, debug = FALSE)
}
\description{
Basic Forward Simulation
}

