\name{msdeTest-package}
\alias{msdeTest-package}
\alias{msdeTest}
\docType{package}
\title{
\packageTitle{msdeTest}
}
\description{
\packageDescription{msdeTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{msdeTest}
\packageIndices{msdeTest}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{msdeTest}

Maintainer: \packageMaintainer{msdeTest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
