% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdeTest.R
\docType{package}
\name{msdeTest}
\alias{msdeTest}
\alias{msdeTest-package}
\title{Simulation and inference for multivariate stochastic differential equation models}
\description{
Simulation and inference for multivariate stochastic differential equation models
}

