% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdeTest.R
\docType{package}
\name{msdeTest2-package}
\alias{msdeTest2}
\alias{msdeTest2-package}
\title{Simulation and inference for multivariate stochastic differential equation models}
\description{
Simulation and inference for multivariate stochastic differential equation models
}

