% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.init.R
\name{sde.init}
\alias{sde.init}
\title{Posterior Sampler Initialization}
\usage{
sde.init(data, dt, k, m, par.index, params, debug = FALSE)
}
\arguments{
\item{data}{An array of data}

\item{dt}{A vector of Delta-t. Must have \code{length(dt) == 1 || length(dt) == nrow(data)-1}}

\item{k}{An integer representing a resolution level (only used if \code{m} is missing)}

\item{m}{An integerrepresenting the number of missing data points per interval}

\item{par.index}{An integer}

\item{params}{A vector of parameters that, if provided, will be included in the output of this function.}

\item{debug}{a boolean (FALSE by default) if set to \code{TRUE}, will cause the function to open a browser mid-call}
}
\value{
a list containing: data, dt, par.index, params (only if \code{params} was passed to sde.init)
}
\description{
initialize posterior sampler for a given resolution level k, \cr
             or m missing data points per interval, i.e. m = 2^k-1. \cr
             currently only supports constant interobservation times.
}
\examples{
# Create the model
hest.model <- sde.make.model(list = hestList, model.name = "hest")

theta <- c(alpha = .1, gamma = 5, beta = .8, sigma = .6, rho = -.7)
Y0 <- c(X = log(100), Z = .1)

# simulate data
N <- 10
burn <- 10
dT <- 1/252

hsim <- sde.sim(model = hest.model, init.data = Y0, params = theta, dt = dT, dt.sim = dT/100,
               N = N, burn = burn, nreps = 1)

k <- 1
par.index <- 1

init <- sde.init(data = hsim$data, dt = dT, k = k, par.index = par.index, params = theta)
}

