% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.make.model.R
\name{sde.make.model2}
\alias{sde.make.model2}
\title{Create sde.model object}
\usage{
sde.make.model2(code, ndims, nparams, data.names, param.names, custom.names,
  cpp.out = TRUE, ..., debug = FALSE)
}
\description{
Create sde.model object
}

