% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.post.R
\name{sde.post2}
\alias{sde.post2}
\title{Basic MCMC sampler for Multivariate SDEs}
\usage{
sde.post2(model, init, init.data, init.params, par.index, dt, nsamples, burn,
  data.out.ind, prior, fixed.params, rw.jump.sd = NULL, update.data = TRUE,
  update.params = TRUE, loglik.out = FALSE, last.miss.out = FALSE,
  verbose = TRUE, debug = FALSE)
}
\description{
Basic MCMC sampler for Multivariate SDEs
}

